+++ src/3rdparty/walkontable/index.js
@@ -106,9 +106,9 @@ function init() {
     offsetColumn: 0,
     height: 200,
     width: 200,
-    rowHeaders: [function (row) {
-      return row + 1
-    }],
+    rowHeaders: function (row, TH) {
+      TH.innerHTML = row + 1;
+    },
     selections: {
       area: {
         className: 'area',